+++ ntex-bytes/src/bytes.rs
@@ -589,6 +589,12 @@ impl From<&Bytes> for Bytes {
     }
 }
 
+impl From<crate::ByteString> for Bytes {
+    fn from(src: crate::ByteString) -> Bytes {
+        src.into_bytes()
+    }
+}
+
 impl From<Vec<u8>> for Bytes {
     /// Convert a `Vec` into a `Bytes`
     fn from(src: Vec<u8>) -> Bytes {