+++ ntex/src/client/h1proto.rs
@@ -1,11 +1,13 @@
 use std::task::{Poll, ready};
 use std::{future::poll_fn, io, io::Write, pin::Pin, task, time::Instant};
 
+use crate::error::{Error, ErrorMapping, with_service};
 use crate::http::body::{Body, BodySize, MessageBody};
 use crate::http::error::PayloadError;
 use crate::http::header::{HOST, HeaderValue};
 use crate::http::{Payload, PayloadStream, ResponseHead, Version, h1};
 use crate::io::{IoBoxed, RecvError};
+use crate::service::cfg::Configuration;
 use crate::time::{Millis, timeout_checked};
 use crate::util::{BufMut, Bytes, BytesMut, Stream};
 
@@ -15,12 +17,27 @@ use super::{ClientCodec, ClientPayloadCodec, ClientRawRequest, pool::Acquired};
 
 pub(super) async fn send_request(
     io: IoBoxed,
+    req: ClientRawRequest,
+    body: Body,
+    created: Instant,
+    timeout: Millis,
+    pool: Option<Acquired>,
+) -> Result<(ResponseHead, Payload), Error<ClientError>> {
+    with_service(
+        io.cfg().ctx().service(),
+        send_request_inner(io, req, body, created, timeout, pool),
+    )
+    .await
+}
+
+async fn send_request_inner(
+    io: IoBoxed,
     mut req: ClientRawRequest,
     body: Body,
     created: Instant,
     timeout: Millis,
     pool: Option<Acquired>,
-) -> Result<(ResponseHead, Payload), ClientError> {
+) -> Result<(ResponseHead, Payload), Error<ClientError>> {
     // set request host header
     if !req.head.headers.contains_key(HOST)
         && let Some(host) = req.head.uri.host()
@@ -42,7 +59,7 @@ pub(super) async fn send_request(
 
     // send request
     let codec = ClientCodec::new(true, io.shared().get());
-    io.send(req.into(), &codec).await?;
+    io.send(req.into(), &codec).await.into_error()?;
 
     log::trace!("http1 request has been sent");
 
@@ -94,17 +111,18 @@ pub(super) async fn send_body(
     mut body: Body,
     io: &IoBoxed,
     codec: &ClientCodec,
-) -> Result<(), ClientError> {
+) -> Result<(), Error<ClientError>> {
     loop {
         if let Some(result) = poll_fn(|cx| body.poll_next_chunk(cx)).await {
-            io.encode(h1::Message::Chunk(Some(result?)), codec)?;
-            io.flush(false).await?;
+            io.encode(h1::Message::Chunk(Some(result.into_error()?)), codec)
+                .into_error()?;
+            io.flush(false).await.into_error()?;
         } else {
-            io.encode(h1::Message::Chunk(None), codec)?;
+            io.encode(h1::Message::Chunk(None), codec).into_error()?;
             break;
         }
     }
-    io.flush(true).await?;
+    io.flush(true).await.into_error()?;
 
     Ok(())
 }
@@ -159,9 +177,10 @@ impl Stream for PlStream {
                         return Poll::Ready(None);
                     }
                 }
-                Err(RecvError::KeepAlive) => {
-                    Err(io::Error::new(io::ErrorKind::TimedOut, "Keep-alive").into())
-                }
+                Err(RecvError::KeepAlive) => Err(PayloadError::from(io::Error::new(
+                    io::ErrorKind::TimedOut,
+                    "Keep-alive",
+                ))),
                 Err(RecvError::WriteBackpressure) => {
                     ready!(this.io.as_ref().unwrap().poll_flush(cx, false))?;
                     continue;